♻️ refactor: ProgressBar 분리
