+++ src/app/quiz/components/QuizButton/type.ts
@@ -1,0 +1,10 @@
+export interface QuizButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  isSubmitted: boolean;
+  isSelected?: boolean;
+  OXType?: 'O' | 'X';
+  imageUrl?: string;
+  percentage?: number;
+  participationLabel?: string;
+  name: string;
+}
